ocr/pyproject.toml
[project]
name = "OCR"
version = "0.1.0"
description = "OCR to Markdown pipeline"
requires-python = ">=3.11"
dependencies = [
    "fpdf2>=2.8.4",
    "pillow>=10.2,<11",
    "python-pptx>=1.0.2",
    "pypdf>=6.0.0",
    "pytesseract>=0.3.13",
    "pyyaml>=6.0.1",
    "aiofiles>=23.2.1",
    "tqdm>=4.67.1,<5",
    "marker-pdf>=1.8.4",
    "magic-pdf[lite]>=1.3.12",
    "olmocr>=0.3.3",
    "torch>=2.5.1",
    "transformers>=4.53.0,<4.54.0",
    "huggingface-hub>=0.24.0",
    "fuzzysearch>=0.7.3",
    "playwright>=1.54.0",
    "ultralytics>=8.3.0",
    "doclayout-yolo>=0.0.4",
    "rapid-table>=2.0.3",
    "pypdfium2>=4.27.0",
    "accelerate>=0.34.2",
    "compressed-tensors>=0.5.1",
    "pymupdf>=1.24.10",
    "numpy>=1.26,<2",
]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
target-version = "py311"
extend-exclude = ["bench"]

[tool.ruff.lint]
extend-select = ["PLR0912"]

[tool.ruff.lint.pylint]
max-branches = 12

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["core*"]

